MOD double to BigDecimal

+++ backend/core/src/main/java/com/scoinone/core/dto/response/order/GetBuyOrderResponseDto.java
@@ -3,14 +3,16 @@ package com.scoinone.core.dto.response.order;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class GetBuyOrderResponseDto {
     private String orderId;
     private String buyerId;
     private String virtualAssetId;
-    private Double quantity;
-    private Double price;
+    private BigDecimal quantity;
+    private BigDecimal price;
     private String status;
     private String tradeTime;
     private String createdAt;
